src/requirements.txt
sounddevice
numpy
torch
torchaudio
pyannote.audio
confluent-kafka
PyYAML